Exit Placement Exercise
Given a graph, with a number of individuals associated to each node, a number of exits, an evacuation capacity of each edge (connecting the nodes to each other and towards the exits) and a maximum evacuation time expressed in iterations, you will have to determine where the exits would be best placed. The number of people evacuated per exit(s) will be the same.